requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backrefs==5.9
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
bracex==2.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.2.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
evaluate==0.4.6
executing==2.2.1
fastapi==0.104.1
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ghp-import==2.1.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
jiter==0.11.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mergedeep==1.3.4
mistune==3.1.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.3
mkdocs-awesome-pages-plugin==2.10.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.23
mkdocs-material-extensions==1.3.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
notebook_shim==0.2.4
numpy==1.26.4
openai==2.6.0
packaging==25.0
paginate==0.5.7
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
pymdown-extensions==10.16.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
pyzmq==27.1.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.27.0
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.2.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.24.0
watchdog==6.0.0
wcmatch==10.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0

transformers
peft
bitsandbytes
accelerate
pyyaml
datasets
mlflow>=2.9.0

